run.py: fix camera smoothing with custom frame range
When start frame is > 0, instead of just rendering a single
frame before the requested range, we render them all using
a very small resolution. This is hacky, but should be pretty
robust w.r.t. renderer internal state.

In the future, this would be better served by a no-op render
call.